widget_api: move message handler into own module
