+++ crates/matrix-sdk/src/widget_api/messages/incoming/mod.rs
@@ -2,7 +2,7 @@ pub mod messages;
 pub mod reply;
 
 pub use self::{
-    messages::{Incoming, SupportedVersions, ApiVersion},
+    messages::{ApiVersion, Incoming, SupportedVersions},
     reply::Reply,
 };
 pub use super::super::Error;
@@ -12,9 +12,8 @@ pub struct Request<C, R> {
     reply: Reply<C, R>,
 }
 
-impl <C, R> Request<C, R> {
+impl<C, R> Request<C, R> {
     pub fn reply(self, response: R) -> Result<(), Error> {
         self.reply.reply(response).map_err(|_| Error::WidgetDied)
     }
 }
-
