+++ crates/matrix-sdk/src/widget_api/messages/mod.rs
@@ -3,6 +3,6 @@ mod message;
 mod outgoing;
 
 pub use self::{
-    incoming::{Incoming, SupportedVersions, ApiVersion},
-    outgoing::{Outgoing, SendMeCapabilities, CapabilitiesUpdated},
+    incoming::{ApiVersion, Incoming, SupportedVersions},
+    outgoing::{CapabilitiesUpdated, Outgoing, SendMeCapabilities},
 };
